+++ src/components/home/HomeFeaturedSection.tsx
@@ -1,12 +1,13 @@
 'use client'
 
+import { PAGE_CONTAINER_CLASS } from '@/constants/page-shell'
 import { HomeFeatureCards } from './HomeFeatureCards'
 import { HomeManInTheHat } from './HomeManInTheHat'
 
 export function HomeFeaturedSection() {
 	return (
 		<section className='py-20 bg-surface-section'>
-			<div className='container mx-auto px-4'>
+			<div className={PAGE_CONTAINER_CLASS}>
 				<HomeManInTheHat />
 				<HomeFeatureCards />
 			</div>